+++ aura/CHANGELOG.md
@@ -19,6 +19,8 @@
 
 - `-Sl` and `-Qil` now work as expected. [#636] [#642]
 - Aura no longer displays the misleading `Please check your input.` message. [#639]
+- File permission issues when building `*-git` packages. [#651] [#634]
+- Bash completions are now possible again for Pacman commands. [#641]
 
 [#647]: https://github.com/fosskers/aura/pull/647
 [#643]: https://github.com/fosskers/aura/issues/643
@@ -26,6 +28,9 @@
 [#642]: https://github.com/fosskers/aura/issues/642
 [#639]: https://github.com/fosskers/aura/issues/639
 [#644]: https://github.com/fosskers/aura/issues/644
+[#651]: https://github.com/fosskers/aura/issues/651
+[#634]: https://github.com/fosskers/aura/issues/634
+[#641]: https://github.com/fosskers/aura/issues/641
 
 ## 3.1.9 (2020-09-11)